requirements.txt
PyQt5==5.15.10
pyinstaller==6.16.0
Pillow>=10.0.0
